CDD: Strongly recommend predictable CPU performance mode in pro audio.
Predictable CPU performance mode maintains the CPU clock at a stable
rate while low-latency audio is active, permitting applications to have
more consistent CPU bandwidth available on short notice.  Without this
mode, if a low-latency audio task is not using high CPU load for a
moment, the governor can decrease the CPU bandwidth.  However, when the
audio task's CPU load increases, the governor may be unable to respond
quickly enough to meet the audio task's requirements, resulting in an
audible glitch. This mode solves that problem and so is recommended for
devices that claim the pro audio feature.

Bug: 32809154
Test: builds OK
Change-Id: I7006af18a614dca841527053b48a83dbf3dd7280