CDD: Add requirement to provide end-to-end IPv6 connectivity.
Direct IPv6 connectivity is standard industry practice on all
Android hardware, simplifies application development, and is
enforced by existing tests. Add a requirement to the CDD to
ensure that apps can rely on it, and any devices or form
factors know that they have to provide it.

Also fix a typo in the section that is being modified.

Bug: 78482227
Test: python make_cdd.py --version 9.0 --branch pi-dev --output /tmp/cdd
Change-Id: Id5606ec179b60f19493b5927c83cdfc411c82203