CDD: Clarification to HDCP requirement for automotive.
Automotive device implementations MUST support HDCP 1.2 or higher for
all external displays connected via a user accessible wired port as
opposed to earlier the requirement was all external displays
connected.

The justification is that, it is unnecessary to specify the video
standard for a permanently installed connection (eg: instrument cluster
or rear seat entertainment screen in a car).


Change-Id: Iafda116268286355a74613b79d9d1ddf514f0f2f
Fixes: 74887951
Test: none