CDD: Clarify default dialer language
The intention of the existing CDD language is to communicate that a
device should allow the user to choose a different "Dialer/Phone"
app to replace the stock dialer/phone app.  This not only means for
placing calls, but also for providing the UI for incoming calls as
well.

We have seen reports in the field from developers creating such an
app that on some devices, the setting will work for most calls, but
sometimes incoming calls will result in the stock dialer UI being
shown instead of the user selected default dialer.  Since this is
an intermittent issue its hard to test with a CTS test.  Setting
the expectation through better requirements language.

Bug: 73368490
Test: Doc update
Change-Id: Ieed7c0c98a131c5749ffb64a93355f942eb02d5c