CDD: Add requirement to provide end-to-end IPv6 connectivity.
Direct IPv6 connectivity is standard industry practice on all
Android hardware, simplifies application development, and is
enforced by existing tests. Add a requirement to the CDD to
ensure that apps can rely on it, and any devices or form
factors know that they have to provide it.

Also fix a typo in the section that is being modified.

Bug: 78482227
Test: python make_cdd.py --version 9.0 --branch pi-dev --output /tmp/cdd
Change-Id: Id8f01bea0c24d03d15c56f95b4742844ab4c6a61