CDD: Requirements for ways to logout accessible from lockscreen
DevicePolicyManager.setLogoutEnabled enables a way to logout (stop current user and switch to user 0), but OEM can decide where to put the button, thus making CTS impossible. e.g. On Pixel, button is on power button menu and lockscreen. Want to include the requirement on CDD to make sure such a button is implemented.

Bug: 73800782
Test: None
Change-Id: Ie48a0ac8431fc0a39603dd996fcb8317394fb8e5