CDD: require timestamps to be accurate to +/- 1 ms
Developers need a reliable and accurate method of obtaining a device's
audio latency so that they can optimize their app's performance for
that device.  As it stands there has been insufficient testing with
the "getTimestamp" method of obtaining audio device latency.  Adding a
requirement to CDD permits the accuracy of getTimestamp to be tested in
CTS or CTS verifier.  The initial strength of requirement is STRONGLY
RECOMMEND.  The corresponding test is tracked separately.

Test: CTS or CTS verifier
Bug: 64140129
Change-Id: If569a2f494534e101587e90afb72dc5ce3250bcb