CDD: Requirements for Heavyweight app mode
Some apps like games have special needs for memory management, since they
tend to be very heavy on memory resources and unable to restore
to their last state (participate in Android’s state saving mechanism).
To preserve games in memory, Android P offers a heavyweight app mode that
will keep an app persistent in memory. There can be only one heavyweight
app designated at a time. If there are multiple heavyweight apps running,
the user will be asked to designate which app should stay in memory. To identify
themselves as a heavyweight app, the app designates a flag in their manifest,
"cantSaveState".  The entire experience is available in AOSP and does not require
modification. Device implementations are required to respect this behavior.

Bug: 78457189
Change-Id: I5397f169923abe530fca17af6a5631adbefeed08
Test: tested in CDD editor