CDD: Requirement for Wi-Fi and BLE scanning.
We have observed that Wi-Fi scanning and BLE scanning is ON for most users
(pre-checked in setup wizard) which means even if they turn off Wi-Fi
and Bluetooth, the phone can still scan for wifi networks and bluetooth devices
because these scan settings are on. To protect user privacy,
device implementations are required to provide user an option to turn Wi-Fi
or BLE scanning on/off.

Bug: 74087421

Test: N/A
Change-Id: I912e1de1f9ccdeeac0c69f70758a1043b47a1b6b