CDD: Telecom Call Redirection
The intention of the CDD language proposed is to ensure that an OEM
supports the android.app.role.CALL_REDIRECTION role and the APIs
surrounding it. Because choosing an app to fill a role is not something
which can be tested easily using an automated CTS test, the CDD
language to enforce the availability of the role is appropriate.

Bug: 124124880
Test: Verified
Change-Id: I7b32d0b841f2d2b3ac94463e458d3376a52cc6a7