CDD: Restrict permitted BUILD.INCREMENTAL and TAGS
Restrictions to the permitted values of BUILD.INCREMENTAL and TAG
are needed to ensure that they do not contain characters that are
not permitted in the build fingerprint or could be confused with
field separators when parsing the build fingerprint.
Permitting arbitary replacement of invalid characters can lead to
ambiguity in the fingerprint.

BUG: 80051029

Change-Id: Ie562c9b6a9a5664e925e9e90453380a7f16c2578