CDD: Add requirement for Wi-Fi Easy Connect
Wi-Fi Easy Connect is a new standard from the Wi-Fi Alliance (WFA).
It is used to provision credentials on devices. The provisioning is
done over a secure channel using separate credentials obtained out-
of-band (e.g. QR code scan). The protocol is a replacement for WPS
(e.g. push-button) which has been determined to be insecure and was
deprecated in Android P.

Android Q supports 2 Easy Connect use-cases:
- Provision credentials on the device from an external source (e.g. AP)
- Provision credentials on IoT devices from the device itself.

Bug: 124063637
Test: manual test with go/markdownr
Change-Id: I0d6a899032b9d2a53e821d0ae70d07e43adfe592