CDD: Add SR requirement to display contextual actions in notifications.
In the javadoc for Notification.Action.Builder#setContextual it is not
clear how contextual actions should be displayed (we don't want to lock
down the implementation/UI unnecessarily). With this CDD insertion we
clarify where contextual actions should be shown.
This only applies to handheld devices since the contextual-actions
feature only targets handheld devices.

BUG: 124111632
Test: Used go/markdownr
Change-Id: Ifccbbfa4b2fa3c6d02aa9fffd7bb49465b4872d2