Revise gamepad requirements
Our gamepad requirements are currently incorrect. For example, we
require all Android devices to handle gamepads, but in the requirements,
it seems that only the devices with "android.hardware.gamepad" do this.

Fix this by splitting a separate section for the gamepad shipped with
the Android device.

The purpose of the feature flag `android.hardware.gamepad` is to indicate
to app developers that a given device was shipped with
an Android-compatible controller in the box. The developer is then
guaranteed that the user will have a gamepad they can use for their
gamepad-only Android application. However, it is still expected that all
Android devices (phone, car, TV, etc.) should have gamepad support,
because a user may connect a supported gamepad to the device at any time.
It's not the expectation that the only devices with gamepad support are
the ones that claim the above feature flag.

Bug: 112265413
Bug: 145056738
Test: visually reviewed the page that opened at http://localhost:8090/
Change-Id: I715300c5c3178225be10f63bb6a493da035da833