CDD: Add section for contacts local account
Standardize local account types in the ContactsProvider.

There is no standard way for applications to refer to local contacts on
Android. It is slightly different on devices from different
manufacturers. So applications  (e.g. contacts backup/restore or
device-to-device contact transfer applications) that want to operate on
local contacts on any Android device must understand the idiosyncrasies
of each manufacturer.

Test: N/A

Bug: 145120250
Change-Id: Ic4818bc7a630a5ec603338c0eafbfcb851ad0584