CDD: Added requirements of file-based on-access verification
The new articles require device implementation to support on-access
verification with trusted certificates, such that for an enabled file,
if a part of the file is tampered with, a read from the tampered part
will fail.

As an example, fs-verity, which is an implementation in Linux kernel
and is used to protect an APK if the APK is installed with a trusted
signature.

Test: check in an MD viewer
Bug: 144365636
Change-Id: Icae88a7cc3e4cdb61cf08cab98ab8adfa2931f77