CDD: Requirements for PIP window size
To respect the activity's minWidth and minHeight when accounting
for the minimal size for PIP.

Applications other than video-based ones may prefer a much more
condensed window when in PIP mode. Previously, the final size
of a PIP window was calculated by the system without the application
being involved.

Summary:
- Separate the requirement for TV from non-TV
- Respect the minimum dimension declared by application as the PIP
window size, if applicable

Bug: 144849584
Test: N/A
Change-Id: Ic6d4793475ed227eaa1b5b1e394dfa4dd2a87250