CDD: strongly recommend kernel stack initialization
The idea is to eliminate bugs related to using uninitialized local
variables in the kernel by force-initializing all the locals. This
includes potential stability bugs as well as information leaks as well
as vulnerabilities related to control flow subversion. Together with
heap initialization, this change is going to mitigate most of the bugs
related to uninitialized memory in the kernel.

Test: None
Bug: 143863382

Signed-off-by: Alexander Potapenko <glider@google.com>
Change-Id: Ia0fe68df775a89c1d49b8d348fd105dcb41ff494