CDD: Strongly recommend kernel heap initialization
The idea is to eliminate bugs related to using uninitialized heap
variables in the kernel by force-initializing all the heap allocations
(page alloc and kmalloc()). This includes potential stability bugs as well
as information leaks as well as vulnerabilities related to control flow
subversion. Together with stack initialization, this change is going to
mitigate most of the bugs related to uninitialized memory in the kernel.

Test: None
Bug: 143931827

Signed-off-by: Alexander Potapenko <glider@google.com>
Change-Id: I3af6f5d8a02fd3895b9c5e125a602e8672478488