CDD: Relaxed Garage Mode requirements on Automotive.
Relaxed the time requirement for automotive products for
the Garage mode functionality as the android automotive ecosystem
is not ready yet. The main Garage Mode goal is to make sure
the platform and associated apps/services are up to date. We
still encourage OEMs to continuously improve their power and
connectivity architectures to better align with the Android
platform needs.

Bug: 155510364
Test: NA
Change-Id: I80dd3a83cdf69dcf29a9406678c96848d162f312