CDD: Explain Vulkan dEQP test requirements
Device implementations that support Vulkan must provide a feature
flag, android.software.vulkan.deqp.level, specifying the version of
the Vulkan dEQP tests that the device claims to pass.  This allows for
flexibility for device makers in the level of Vulkan dEQP tests that
are supported, and allows apps to query the level of support that is
provided.  This CDD change sets out the requirements related to this
new feature flag.

Bug: 148619801
Change-Id: I59aec5fed4bca33473f184a362d34a7b7cb450b3