CDD: Handheld requirements for device controls.
Add definitions to support Device Controls. Device
Controls are an affordance within the Power Menu that allows users to
remotely control their home devices such as lights and thermostats,
as well as other IoT devices. SystemUI provides a default
implementation of this feature for AOSP.

The user experience for home control (IoT control) is fragmented
and not integrated across multiple vendor-supplied applications.
Adding in requirements to create a standard surface for the
individual controls provided by app developers to show on.

Bug: 145990227
Test: go/marked
Change-Id: Ic5cdb781683f3610f9600d40ea1f8fc568cb52ae