CDD: Intents visible to Instant Apps.
An Android Instant App is a small version of your app
that runs without installation. Since Instant Apps cannot see
the other applications on the device and rely on pickers and
other system-defined components to do work, the request is
to expose certain intents and make them available to instant apps
i.e an instant application can launch this intent and there will
be a handler for it.

Bug: 147893121
Change-Id: Ied14e0e4753a3ddf2150b32387047b93ad9d23db