CDD: Intents Classification and Clarification
Intents have been classified as application intents and
broadcast intents.
Application intents have been listed for each form factor.
Removed the terminology of Core intents and called it common
application intents to be more inline with the developer pages.

Also renamed section "3.2.3.5 default app settings" to
"conditional application intents" and moved in the conditional
application intents in that section.

The goal is to provide clarity to OEM's and developers on the
list of intents to expect an activity/handler.

Change-Id: I4416c2b06b7845581e701f8137e7d870d4749938
BUG: 148181180