CDD: Remove .ota as MIDI file extension
In Android 11 (R), the .ota file extension is no longer for MIDI,
but is for its official MIME mapping, to support Android local
system updates.

This change matches the CDD with what the plaform code is already
doing in Android 11, and what CTS is enforcing.

See:
https://android-review.googlesource.com/c/platform/libcore/+/999958
for more detail on platform code change.

Bug: 181140216
Test: None
Change-Id: I56abdb202516192710f9c62bf2a7e15efdd6002e